Factor_Class
The Factor class is an encapsulated class . Each factor object encapsulates a factor value. The initial state of the factor object is Active. The state transitions to inactive if the user's previous Input equals their second input. The state of the object can only be reset if the state of the object is inactive.
Language: C#
Platform : Visual Studio